pyproject.toml
[project]
name = "regulatory-intelligence-ai"
version = "0.1.0"
description = "AI-powered regulatory change monitoring system for RBI/SEBI using LangChain and Groq"
readme = "README.md"
requires-python = ">=3.11"

authors = [
    { name = "Ranganathan J", email = "your-email@example.com" }
]

dependencies = [
    # -----------------------------
    # Core
    # -----------------------------
    "python-dotenv>=1.0.1",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "PyYAML>=6.0.1",
    "loguru>=0.7.2",

    # -----------------------------
    # Web Scraping
    # -----------------------------
    "requests>=2.32",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "feedparser>=6.0",

    # -----------------------------
    # Data Handling
    # -----------------------------
    "pandas>=2.2",
    "openpyxl>=3.1",

    # -----------------------------
    # LangChain + LLM
    # -----------------------------
    "langchain>=0.2",
    "langchain-core>=0.2",
    "langchain-community>=0.2",
    "langchain-groq>=0.1",

    # -----------------------------
    # Notifications
    # -----------------------------
    "msal>=1.28",          # Outlook / Microsoft Graph
    "requests-oauthlib>=2.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "black>=24.4",
    "ruff>=0.5",
    "mypy>=1.10"
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "src/crawler"]
